feat: implement core library, playback, and lyrics service architecture with MVVM viewmodels

+++ Octave.Core/Models/DomainModels.cs
@@ -40,7 +40,7 @@ public record Track(
     int Year,
     DateTime DateAdded,
     string Genre = "",      // Trailing default so existing constructors are unaffected
-    double ReplayGain = 0.0 // Applied in playback to normalize loudness
+    float ReplayGain = 0.0f // Applied in playback to normalize loudness
 );
 
 public record Playlist(
